Crate wuff

Pure Rust WOFF and WOFF2 decoder

Modules

brotli 🔒

Built-in WOFF2 Brotli decompression, backed by brotli-decompressor.

decompress_woff1 🔒

decompress_woff2 🔒

error 🔒

table_tags 🔒

Font table tags

variable_length 🔒

Helper functions for woff2 variable length types: 255UInt16 and UIntBase128

woff 🔒

Macros

Round4 🔒

Structs

Point 🔒

Enums

WuffErr

Constants

GLYF 🔒

HEAD 🔒

HHEA 🔒

HMTX 🔒

LOCA 🔒

Functions

compute_checksum 🔒

Compute checksum over size bytes of buf

decompress_woff2

Decompress a WOFF2 file using the built-in brotli decompressor

decompress_woff1_with_custom_z

Decompress a WOFF1 file using a custom gzip decompressor passed as a closure

decompress_woff2_with_custom_brotli

Decompress a WOFF2 file using a custom brotli decompressor passed as a closure

write_table_directory_header 🔒

Writes an OpenType table directory