wuff::woff

Module glyf_decoder

Structs

GlyfAndLocaData

GlyfDecoder

Constants

COMPOSITE_GLYPH_BEGIN 🔒

DEFAULT_GLYPH_BUF_SIZE 🔒

END_PTS_OF_CONTOURS_OFFSET 🔒

FLAG_ARG_1_AND_2_ARE_WORDS 🔒

FLAG_MORE_COMPONENTS 🔒

FLAG_OVERLAP_SIMPLE_BITMAP 🔒

FLAG_WE_HAVE_AN_X_AND_Y_SCALE 🔒

FLAG_WE_HAVE_A_SCALE 🔒

FLAG_WE_HAVE_A_TWO_BY_TWO 🔒

FLAG_WE_HAVE_INSTRUCTIONS 🔒

GLYF_ON_CURVE 🔒

GLYF_REPEAT 🔒

GLYF_THIS_X_IS_SAME 🔒

GLYF_THIS_Y_IS_SAME 🔒

GLYF_X_SHORT 🔒

GLYF_Y_SHORT 🔒

NUM_SUB_STREAMS 🔒

OVERLAP_SIMPLE 🔒

Functions

compute_size_of_composite 🔒

decode_triplet 🔒

generate_loca_table 🔒

Generate a loca table given a slice of loca offsets and an index format

tranform_glyf_table 🔒

Decode a WOFF2 transformed glyf table

write_bbox 🔒

Compute the bounding box of the coordinates, and store into a glyf buffer. A precondition is that there are at least 10 bytes available. dst should point to the beginning of a ‘glyf’ record.

write_glyph_points 🔒