wuff::woff::headers

Struct WoffHeader

pub struct WoffHeader {

    pub woff_version: WoffVersion,
    pub signature: Tag,
    pub flavor: Tag,
    pub length: u32,
    pub num_tables: u16,
    pub reserved: u16,
    pub total_sfnt_size: u32,
    pub total_compressed_size: u32,
    pub major_version: u16,
    pub minor_version: u16,
    pub meta_offset: u32,
    pub meta_length: u32,
    pub meta_orig_length: u32,
    pub priv_offset: u32,
    pub priv_length: u32,

}

WOFF header that can represent either a WOFF1 or WOFF2 header

https://www.w3.org/TR/WOFF2/#woff20Header

Fields
woff_version: WoffVersionsignature: Tag

b“woFF“ or b“wOF2“

flavor: Tag

The “sfnt version” of the input font.

length: u32

Total size of the WOFF file.

num_tables: u16

Number of entries in directory of font tables.

reserved: u16

Reserved; set to 0.

total_sfnt_size: u32

Total size needed for the uncompressed font data, including the sfnt header, directory, and font tables (including padding).

total_compressed_size: u32

(WOFF2 only) Total length of the compressed data block.

major_version: u16

Major version of the WOFF file.

minor_version: u16

Minor version of the WOFF file.

meta_offset: u32

Offset to metadata block, from beginning of WOFF file.

meta_length: u32

Length of compressed metadata block.

meta_orig_length: u32

Uncompressed size of metadata block.

priv_offset: u32

Offset to private data block, from beginning of WOFF file.

priv_length: u32

Length of private data block.

Implementations

impl WoffHeader

pub fn parse(input: &mut impl Buf) -> Result<Self, WuffErr>

pub fn is_collection(&self) -> bool